Fallback to `os.homedir()` on Windows
The home directory will always be the correct one on Windows on
elevation, so we fallback to the default way of getting the home
directory in this case.

We could have used `os.homedir()`, which was added in NodeJS v4, however
we use a shim. `os-homedir` to cover older versions as well.

See https://github.com/sindresorhus/os-homedir
Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>